+++ geometry/nxeasing.h
@@ -38,7 +38,7 @@ public:
 public:
     inline void setType(quint16 type)           { easing.setType((QEasingCurve::Type)type); }
     inline quint16 getType() const              { return easing.type(); }
-    inline qreal getValue(qreal progress) const { return qBound(0., easing.valueForProgress(progress), 1.); }
+    inline qreal getValue(qreal progress) const { return qBound(qreal(0.), easing.valueForProgress(progress), qreal(1.)); }
 };
 
 #endif // NXEASING_H
